backend/requirements.txt
flask
flask-cors
sqlalchemy
psycopg2-binary
bcrypt
gunicorn